🚚 :: Board 폴더 분할
